Error out if empty username.
The player_locations sensor is the only place where the empty username
actually matters (the empty string is otherwise perfectly permissible as
a table key -- it's when we try to use it as a GUI element name that it
fails).

I could technically have provided a simple prefix or suffix to help, but
I'd rather just fail outright in this one rare case.

Fixes #15 and bumps the version to 0.4.1.