Work around 0.15.22 bug with disappearing GUIs
As per the comment in the source, if a mod creates a GUI element whose
name is both:
  - longer than 4 characters, and
  - less than 4 characters longer than the mod name,
the element would be considered to belong to a mod that isn't present
and get removed.

We fixed this earlier in its guise as #71, which is how the bug appears
in single-player, but in multiplayer the removal runs only client-side
on joining, causing an instant desync for the joining player.

My workaround is to rename the root element to evogui_root and hope
(suspect, believe) that the remaining GUI elements all have longer names
or won't cause instant breakage (so far, this seems to be holding).

On the other hand, updating just EvoGUI while still on Factorio 0.15.22
needs to be a two-step process: save the game once, *then* reload it
(start server, stop server, start server), as otherwise the client will
cause the same desync again. It's possible just waiting until the first
server autosave may also fix it, but it's quicker this way.

Closes #74.